backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ─── Pydantic v2 ──────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── JWT (RS256) ──────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
cryptography==42.0.5

# ─── Rate Limiting ────────────────────────────────────────────────────────────
slowapi==0.1.9
limits==3.12.0

# ─── Async HTTP Client (Proxy) ────────────────────────────────────────────────
httpx==0.27.0

# ─── Structured Logging ───────────────────────────────────────────────────────
structlog==24.1.0

# ─── CORS / Security Headers ──────────────────────────────────────────────────
starlette==0.37.2

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0          # also used as async test client

# ─── Dev Utilities ────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Database ─────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.30
aiosqlite==0.20.0

# ─── Hashing ──────────────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
